Support Upgrades
Changing configuration isn't easy, as machine templates and kubeadm
configurations are immutable, so we need to spot changes and rename the
resources and references accordingly.  Tweak the schema so at least one
worker pool is required, and also -- now I know it exists -- enhance
the schema to describe content type for base64 encoded fields to improve
documentation.