Opinionated tweaks (#13)
* Autoscaling and Other Tweaks

Adds in support for basic cluster-autoscaler limits on machine
deployments, then scheduler hints for CPU/memory/GPU to enable scale
from zero support, along with the requisite node labels.  Workload
clusters are split into pools so you can have different classes of
machine in the same heterogenous cluster.  Added additional support at
the API to allow IP prefix whitelists and X.509 SANs to be specified.
Finally, this is all backed up with a schema to validate input, and --
utimately -- allow type generation to aid in configuring this beast.

* Add Node Labels

Adds in the well know "region" label for scheduling, and the "zone"
which can now be specified per pool.  A label is implicitly added per
pool to uniquely identify it, and also the user can now apply custom
labels e.g. those that span pools.

* Support Upgrades

Changing configuration isn't easy, as machine templates and kubeadm
configurations are immutable, so we need to spot changes and rename the
resources and references accordingly.  Tweak the schema so at least one
worker pool is required, and also -- now I know it exists -- enhance
the schema to describe content type for base64 encoded fields to improve
documentation.

* Add File Support

Primarily for non-user facing things like dropping licences and other
configuration on the host for use by 3rd party controllers.  Technically
this could be done with a daemonset.

* Support Staged Upgrades

Worker pools and the control plane need to be independently
imaged/versioned in order to support lower-risk upgrades.  It also has
the side effect of making descriminator determination easier and less
error prone.

* Tweak Node Labels

There is no ambiguity here, so just go with labels rather than
nodeLabels.

* Enable Control Plane Service Listeners

Aids in being able to monitor things if they listen on all interfaces,
not just localhost.

* Move to External Cloud Provider

For this is the future.

* Add AZs for Cinder

Not using this may end in disaster if machine AZs are translated into
AZs that don't exist for the storage provider.