Batch Update
Update the CAPI charts to use the most recent release, modifying the
generator to change the image pull policy (which is whack and doesn't
work on UK trains...).  Hopefully caching here will improve local
performance quite a bit.  As for the cluster, revert the zoning changes
as they should get filled in, make SSH keys optional, and workload pools
as having to handle the "default" specially is a pain further up the
stack.