Make Clusters Kinda Mutable (#32)
Allow the firewall rules to be updated and thus trigger a rename of the
cluster resource (which is immutable).  Everything else is immutable
still.  In future, if the resource changes outside of user input, we'll
need a way to trigger a rename.  Perhaps a manually defined revision or
something that is bumped by the developer.