Handle CAPO Brokeness (#34)
Adding in the control plane failure domains caused upgrade failures
because the osc resource is immutable (with the exception of the
firewall rules).  We tried to just change the name and the reference as
per osmts etc.  This however failed quite dramatically because CAPO
decides to get a new IP address for the API, which will break everything
due to contact with the cluster vanishing.  However while it's doing
this, it's constantly allocating floating IPs and failing to apply them
due to one already existing.  So this reverts the original change that
led to this mess.