Add Autoscaler Side-Chart
While the community want this functionality (justifiably) in the main
chart, as that's the thing that brings in the dependency, the
maintainers (justifiably) don't want platform specific stuff in there.
So rather than battle against the current, or have thousands of users
fork and maintain the required bit, provide them in a centralized chart.
That way you're not worried about forks, just this and the main branch.